Fix reported JNI issues in file-events
